intensity type changed to float, normalized to [0,1] WARNING normalizing in getIntensity.

+++ plugin/src/mosaic/region_competition/ContourParticle.java
@@ -7,7 +7,7 @@ import java.util.List;
 public class ContourParticle 
 {
 	int label = 0;				// absLabel
-	int intensity = 0;
+	float intensity = 0.0f;
 	int candidateLabel = 0;
 	double energyDifference = 0;
